sigue la refactorizacion + acomodados actividades y fragmentos en paquetes
