+++ goodjob/app/src/main/AndroidManifest.xml
@@ -12,19 +12,18 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SolicitudEmpresaResultadoActivity" />
-        <activity android:name=".SolicitudEmpresaActivity" />
-        <activity android:name=".PublicarActividadActivity" />
+        <activity android:name=".activities.SolicitudEmpresaResultadoActivity" />
+        <activity android:name=".activities.SolicitudEmpresaActivity" />
+        <activity android:name=".activities.PublicarActividadActivity" />
         <activity android:name=".SuscriptionActivity" />
-        <activity android:name=".FormRegisterUserActivity" />
-        <activity android:name=".FormRegisterCompanyActivity" />
+        <activity android:name=".activities.FormRegisterUserActivity" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".activities.LoginActivity"
             android:theme="@style/AppTheme.FullScreem" />
-        <activity android:name=".SelectedUserActivity" />
-        <activity android:name=".DetailsAndApplyActivity" />
+        <activity android:name=".activities.SelectedUserActivity" />
+        <activity android:name=".activities.DetailsAndApplyActivity" />
         <activity
-            android:name=".SplashActivity"
+            android:name=".activities.SplashActivity"
             android:theme="@style/SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -33,7 +32,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:label="@string/app_name" />
 
         <meta-data
