+++ goodjob/app/src/main/java/com/example/goodjob/activities/SelectedUserActivity.java
@@ -1,10 +1,12 @@
-package com.example.goodjob;
+package com.example.goodjob.activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
+
+import com.example.goodjob.R;
 
 public class SelectedUserActivity extends AppCompatActivity {
     private TextView tvuser, tvcompany;
