+++ goodjob/app/src/main/java/com/example/goodjob/activities/SolicitudEmpresaActivity.java
@@ -1,7 +1,6 @@
-package com.example.goodjob;
+package com.example.goodjob.activities;
 
 import android.content.Intent;
-import android.support.design.widget.Snackbar;
 import android.support.design.widget.TextInputLayout;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
@@ -20,6 +19,7 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
+import com.example.goodjob.R;
 import com.example.goodjob.classes.ValidSession;
 
 import java.util.HashMap;
