+++ goodjob/app/src/main/java/com/example/goodjob/activities/SplashActivity.java
@@ -1,4 +1,4 @@
-package com.example.goodjob;
+package com.example.goodjob.activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
