+++ goodjob/app/src/main/java/com/example/goodjob/fragments/DetalleMiPublicacionFragment.java
@@ -1,4 +1,4 @@
-package com.example.goodjob;
+package com.example.goodjob.fragments;
 
 
 import android.os.Bundle;
@@ -16,6 +16,7 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
+import com.example.goodjob.R;
 import com.example.goodjob.adapter.UsuarioPostulanteAdapter;
 import com.example.goodjob.classes.ActividadAceptada;
 import com.example.goodjob.classes.UsuarioPostulante;
