+++ goodjob/app/src/main/java/com/example/goodjob/fragments/EstadoMisActividadesFragment.java
@@ -1,4 +1,4 @@
-package com.example.goodjob;
+package com.example.goodjob.fragments;
 
 
 import android.os.Bundle;
@@ -15,6 +15,7 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
+import com.example.goodjob.R;
 import com.example.goodjob.adapter.EstadoMisActividadesAdapter;
 import com.example.goodjob.classes.EstadoMisActividadesResponse;
 import com.example.goodjob.classes.ValidSession;
