+++ goodjob/app/src/main/res/layout/activity_details_and_apply.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="#FFFFCC"
     android:orientation="vertical"
-    tools:context=".DetailsAndApplyActivity">
+    tools:context=".activities.DetailsAndApplyActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
