--- goodjob/app/src/main/res/layout/activity_form_register_company.xml
@@ -1,99 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".FormRegisterCompanyActivity"
-    android:gravity="center">
-
-    <TextView
-        android:id="@+id/textView8"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/Form_Name_Company"
-        android:textSize="16dp"/>
-
-    <EditText
-        android:id="@+id/txtrazonsocialR"
-        android:layout_width="280dp"
-        android:layout_height="wrap_content"
-        android:ems="10"
-        android:inputType="textPersonName" />
-
-    <TextView
-        android:id="@+id/textView9"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/Form_RUC_Company"
-        android:textSize="16dp"/>
-
-    <EditText
-        android:id="@+id/txtrucR"
-        android:layout_width="280dp"
-        android:layout_height="wrap_content"
-        android:ems="10"
-        android:inputType="number" />
-
-    <TextView
-        android:id="@+id/textView10"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/Form_Area_Company"
-        android:textSize="16dp"/>
-
-    <Spinner
-        android:id="@+id/spinner"
-        android:layout_width="250dp"
-        android:layout_height="40dp" />
-
-    <TextView
-        android:id="@+id/textView11"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/Form_cellphone_Company"
-        android:textSize="16dp"/>
-
-    <EditText
-        android:id="@+id/txtcelularRC"
-        android:layout_width="200dp"
-        android:layout_height="wrap_content"
-        android:ems="10"
-        android:inputType="phone" />
-
-    <TextView
-        android:id="@+id/textView12"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/Form_Pass_Company"
-        android:textSize="16dp"/>
-
-    <EditText
-        android:id="@+id/txtpassRC"
-        android:layout_width="280dp"
-        android:layout_height="wrap_content"
-        android:ems="10"
-        android:inputType="textPassword" />
-
-    <EditText
-        android:id="@+id/txtpassRCdos"
-        android:layout_width="280dp"
-        android:layout_height="wrap_content"
-        android:ems="10"
-        android:inputType="textPassword"
-        android:layout_marginTop="15dp"/>
-
-    <Button
-        android:id="@+id/btnRegistrar"
-        android:layout_width="150dp"
-        android:layout_height="wrap_content"
-        android:text="@string/Form_btn_Register" />
-
-    <Button
-        android:id="@+id/btnCancelar"
-        android:layout_width="150dp"
-        android:layout_height="wrap_content"
-        android:text="@string/Form_btn_Noregister" />
-
-</LinearLayout>
