+++ goodjob/app/src/main/res/layout/activity_form_register_user.xml
@@ -7,7 +7,7 @@
     android:gravity="center_horizontal"
     android:paddingTop="16dp"
     android:orientation="vertical"
-    tools:context=".FormRegisterUserActivity">
+    tools:context=".activities.FormRegisterUserActivity">
 
     <ImageView
         android:layout_width="match_parent"
