+++ goodjob/app/src/main/res/layout/activity_publicar_actividad.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PublicarActividadActivity">
+    tools:context=".activities.PublicarActividadActivity">
 
     <View
         android:layout_width="match_parent"
