+++ goodjob/app/src/main/res/layout/activity_selected_user.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".SelectedUserActivity"
+    tools:context=".activities.SelectedUserActivity"
     android:background="@drawable/fondologin">
 
     <TextView
