+++ goodjob/app/src/main/res/layout/activity_solicitud_empresa.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".SolicitudEmpresaActivity">
+    tools:context=".activities.SolicitudEmpresaActivity">
 
     <TextView
         android:layout_width="match_parent"
