+++ goodjob/app/src/main/res/layout/activity_solicitud_empresa_resultado.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="@dimen/default_padding_left"
-    tools:context=".SolicitudEmpresaResultadoActivity">
+    tools:context=".activities.SolicitudEmpresaResultadoActivity">
 
     <TextView
         android:layout_width="match_parent"
