+++ goodjob/app/src/main/res/layout/fragment_actividad_aceptada.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ActividadAceptadaFragment"
+    tools:context=".fragments.ActividadAceptadaFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
