+++ goodjob/app/src/main/res/layout/fragment_detalle_mi_publicacion.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".DetalleMiPublicacionFragment"
+    tools:context=".fragments.DetalleMiPublicacionFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
