+++ goodjob/app/src/main/res/layout/fragment_estado_mis_actividades.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EstadoMisActividadesFragment"
+    tools:context=".fragments.EstadoMisActividadesFragment"
     android:background="#EE46BEB6">
 
     <android.support.v7.widget.RecyclerView
