+++ goodjob/app/src/main/res/layout/fragment_estado_mis_publicaciones.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EstadoMisPublicacionesFragment">
+    tools:context=".fragments.EstadoMisPublicacionesFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/rvMisPublicaciones"
