+++ goodjob/app/src/main/res/layout/fragment_pre_my_activites.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PreMyActivitesFragment" >
+    tools:context=".fragments.PreMyActivitesFragment" >
 
     <Button
         android:id="@+id/btnMisPublicaciones"
